Detect embedded webview on iPadOS.
iPadOS identifies as Mac OS, but ua-parser-js does the extra detection
to determine if it's an iPad. Use the device model to detect when we're
in an embedded webview on iPadOS.

Also removes the version check for best effort webkit detection as
iPadOS sends version 10.15 and checking RTCRtpTransceiver should be
sufficient without the version check.