Fix regression in Chrome 143 due to changes in RTP extensions handling
# Details

- Fixes #353
- Tested in Chrome 143.
- Tested in Opera 122 which uses Chromium version:138.0.7204.251.
- Tested in Firefox 144.0.2.
- Tested in Safari 18.6.

## Rationale

Chrome 143 complains when we disable the first m= section in the sending PeerConnection. It complains because its locally generated offer contains simulcast attributes but the SDP answer mediasoup-client generates doesn't contain the corresponding RTP extensions for simulcast (RID). And since we are not closing the transceiver (because it's the first one so it would disconnect ICE), we only "disable" it which means that the response has "a=inactive" but the lack of RID extension makes Chrome 143 throw an error because the offer requests simulcast.
Solution is to NOT remove RTP extension attributes from the disabled m= section in the remote SDP answer, and only do it when closing the transceiver (which involves remote port set to 0).